clean up the amie page and move out from the main mathapp.py
